Validation code for all fields. BS4 for XML parsing. HamQTH and HamDB for additional lookups.

+++ WFD_Curses.py
@@ -14,9 +14,8 @@ COLOR_YELLOW	Yellow
 import os
 import logging
 from pathlib import Path
-
-confFile = os.path.expanduser("~") + "/.config/wfd.conf"
-
+from bs4 import BeautifulSoup
+ 
 if Path("./debug").exists():
     logging.basicConfig(
         filename="debug.log",
@@ -28,54 +27,86 @@ if Path("./debug").exists():
     logging.debug("Debug started")
 
 cloudLogOn = False
-qrzOn = False
+qrzsession = False
+hamdbOn = False
+hamqthSession = False
 
 try:
     import json
     import requests
 
+    confFile = os.path.expanduser("~") + "/.config/wfd.ini"
+
     if os.path.exists(confFile):
         fd = open(confFile)
         try:
             confData = json.load(fd)
-            cloudlogapi = confData["cloudlog"]["apikey"]
-            cloudlogurl = confData["cloudlog"]["url"]
-            qrzurl = confData["qrz"]["url"]
-            qrzname = confData["qrz"]["username"]
-            qrzpass = confData["qrz"]["password"]
         except Exception as e:
             logging.debug(f"{e}")
         fd.close()
-    else:
-        cloudlogapi = "cl12345678901234567890"
-        cloudlogurl = "http://www.yoururl.com/Cloudlog/index.php/api"
-        qrzurl = "http://xmldata.qrz.com/xml/"
-        qrzname = "w1aw"
-        qrzpass = "secret"
+    elif os.path.exists("/usr/local/etc/wfd.ini"):
+        fd = open("/usr/local/etc/wfd.ini")
+        try:
+            confData = json.load(fd)
+        except Exception as e:
+            logging.debug(f"{e}")
+        fd.close()
+    elif os.path.exists("/etc/wfd.ini"):
+        fd = open("/etc/wfd.ini")
+        try:
+            confData = json.load(fd)
+        except Exception as e:
+            logging.debug(f"{e}")
+        fd.close()
 
-    payload = {"username": qrzname, "password": qrzpass}
-    r = requests.get(qrzurl, params=payload, timeout=1.0)
+    if confData["cloudlog"]["enable"].lower() == "yes":
 
-    if r.status_code == 200 and r.text.find("<Key>") > 0:
-        qrzsession = r.text[r.text.find("<Key>") + 5 : r.text.find("</Key>")]
-        qrzOn = True
-    else:
-        qrzsession = False
-    if r.status_code == 200 and r.text.find("<Error>") > 0:
-        errorText = r.text[r.text.find("<Error>") + 7 : r.text.find("</Error>")]
-        logging.debug(f"QRZ Error: {errorText}")
+        cloudlogapi = confData["cloudlog"]["apikey"]
+        cloudlogurl = confData["cloudlog"]["url"]
 
-    payload = "/validate/key=" + cloudlogapi
-    r = requests.get(cloudlogurl + payload)
+        payload = "/validate/key=" + cloudlogapi
+        r = requests.get(cloudlogurl + payload)
 
-    if r.status_code == 200 or r.status_code == 400:
-        cloudLogOn = True
+        if r.status_code == 200 or r.status_code == 400:
+            cloudLogOn = True
+
+    if confData['qrz']['enable'].lower() == "yes":
+
+        qrzurl = confData["qrz"]["url"]
+        qrzname = confData["qrz"]["username"]
+        qrzpass = confData["qrz"]["password"]
+
+        payload = {"username": qrzname, "password": qrzpass}
+        r = requests.get(qrzurl, params=payload, timeout=1.0)
+
+        if r.status_code == 200:
+            xmlData = BeautifulSoup(r.text, "xml")
+            if xmlData.QRZDatabase.Session.Key.string:
+                qrzsession = xmlData.QRZDatabase.Session.Key.string
+        else:
+            logging.debug(f"{xmldata.QRZDatabase.Session.Error.string}")
+            qrzsession = False
+
+    if confData['hamdb']['enable'].lower() == "yes":
+        hamdbOn = True
+
+    if confData['hamqth']['enable'].lower() == "yes":
+        payload = "?u=" + confData['hamqth']['username'] + "&p=" + confData['hamqth']['password']
+        r = requests.get(confData['hamqth']['url'] + payload)
+        if r.status_code == 200:
+            xmlData = BeautifulSoup(r.text, "xml")
+            hamqthSession = xmlData.HamQTH.session.session_id.string
+        else:
+            hamqthSession = False
 
 except:
     cloudlogapi = False
     cloudlogurl = False
+    cloudLogOn = False
+    HamDbOn = False
     qrz = False
     qrzsession = False
+    hamqthSession = False
 
 import curses
 import time
@@ -137,7 +168,8 @@ dfreqCW = {
     "222": "223.500",
     "432": "425.000",
 }
-modes = ("PH", "CW", "DI")
+
+validSections = [ "CT", "RI", "EMA", "VT", "ME", "WMA", "NH", "ENY", "NN", "NLI", "SNJ", "NNJ", "WNY", "DE",  "MDC", "EPA", "EPA", "AL", "SC", "GA", "SFL", "KY", "TN", "NC", "VA", "NFL", "VI", "PR", "WCF", "AR", "NTX", "LA", "OK", "MS", "STX", "NM", "WTX", "EB", "SCV", "LAX", "SDG", "ORG", "SF", "PAC", "SJV", "SB", "SV", "AK", "NV", "AZ", "OR", "EWA", "UT", "ID", "WWA", "MT", "WY", "MI", "WV", "OH", "IL", "WI", "IN", "CO", "MO", "IA", "ND", "KS", "NE", "MN",  "SD", "AB", "NT", "BC", "ONE", "GTA", "ONN", "MAR", "ONS", "MB", "QC", "NL", "SK", "PE" ]     
 
 mycall = "YOURCALL"
 myclass = "CLASS"
@@ -271,7 +303,7 @@ def sendRadio(cmd, arg):
                 except:
                     rigonline == False
             else:
-                setStatusMsg("Invalid frequency specified")
+                setStatusMsg("Specify frequency in Hz")
         elif cmd == "M":
             rigCmd = bytes(cmd + " " + arg + " 0\n", "utf-8")
             try:
@@ -408,19 +440,22 @@ def log_contact(logme):
 
 
 def delete_contact(contact):
-    try:
-        with sqlite3.connect(database) as conn:
-            sql = f"delete from contacts where id={int(contact)}"
-            cur = conn.cursor()
-            cur.execute(sql)
-            conn.commit()
-    except Error as e:
-        logging.debug(f"delete_contact: {e}")
-        displayinfo(e)
-    workedSections()
-    sections()
-    stats()
-    logwindow()
+    if contact:
+        try:
+            with sqlite3.connect(database) as conn:
+                sql = f"delete from contacts where id={int(contact)}"
+                cur = conn.cursor()
+                cur.execute(sql)
+                conn.commit()
+        except Error as e:
+            logging.debug(f"delete_contact: {e}")
+            displayinfo(e)
+        workedSections()
+        sections()
+        stats()
+        logwindow()
+    else:
+        setStatusMsg("Must specify a contact number")
 
 
 def change_contact(qso):
@@ -433,7 +468,6 @@ def change_contact(qso):
     except Error as e:
         logging.debug(f"change_contact: {e}")
         displayinfo(e)
-
 
 def readSections():
     try:
@@ -674,27 +708,10 @@ def adif():
                     payload = {"s": qrzsession, "callsign": hiscall}
                     r = requests.get(qrzurl, params=payload, timeout=3.0)
                     if r.status_code == 200:
-                        if r.text.find("<grid>") > 0:
-                            grid = r.text[
-                                r.text.find("<grid>") + 6 : r.text.find("</grid>")
-                            ]
-                        if r.text.find("<fname>") > 0:
-                            name = r.text[
-                                r.text.find("<fname>") + 7 : r.text.find("</fname>")
-                            ]
-                        if r.text.find("<name>") > 0:
-                            if not name:
-                                name = r.text[
-                                    r.text.find("<name>") + 6 : r.text.find("</name>")
-                                ]
-                            else:
-                                name += (
-                                    " "
-                                    + r.text[
-                                        r.text.find("<name>")
-                                        + 6 : r.text.find("</name>")
-                                    ]
-                                )
+                        xmlData = BeautifulSoup(r.text, "xml")
+                        if xmlData.QRZDatabase.Callsign.grid:
+                            grid = xmlData.QRZDatabase.Callsign.grid.string
+                        name = xmlData.QRZDatabase.Callsign.fname.string + " " + xmlData.QRZDatabase.Callsign.name.string
             except:
                 pass
             print(
@@ -755,7 +772,7 @@ def adif():
 
 
 def postcloudlog():
-    global qrzsession
+    global confData, hamdbOn, hamqthOn, qrzsession
     if not cloudlogapi:
         return
     conn = sqlite3.connect(database)
@@ -765,6 +782,21 @@ def postcloudlog():
     conn.close()
     logid, hiscall, hisclass, hissection, datetime, band, mode, power = q
     grid = False
+    if hamqthSession:
+        payload = "?id=" + hamqthSession + "&callsign=" + hiscall + "&prg=" + confData['hamqth']['appname']
+        r = requests.get(confData['hamqth']['url'] + payload)
+        if r.status_code == 200:
+            xmlData = BeautifulSoup(r.text, "xml")
+            grid = xmlData.HamQTH.search.grid.string
+            if len(grid) < 4 or len(grid) > 6:
+                grid = ""
+    if hamdbOn:
+        payload = hiscall + "/xml/" + confData['hamdb']['appname']
+        r = requests.get(confData['hamdb']['url'] + "/" + payload)
+        if r.status_code == 200:
+            grid = r.text[r.text.find("<grid>") + 6 : r.text.find("</grid>")]
+        if len(grid) > 6:
+            grid = ""
     if qrzsession:
         payload = {"s": qrzsession, "callsign": hiscall}
         r = requests.get(qrzurl, params=payload, timeout=1.0)
@@ -1229,9 +1261,12 @@ def statusline():
     if len(strband) < 4:
         strband += " "
 
-    stdscr.addstr(20, 36, "QRZ   CloudLog  ")
-    stdscr.addstr(20, 40, YorN(qrzOn), highlightBonus(qrzOn))
-    stdscr.addstr(20, 51, YorN(cloudLogOn), highlightBonus(cloudLogOn))
+    stdscr.addstr(20, 35, " HamDB   HamQTH   ")
+    stdscr.addstr(20, 42, YorN(hamdbOn), highlightBonus(hamdbOn))
+    stdscr.addstr(20, 51, YorN(hamqthSession), highlightBonus(hamqthSession))
+    stdscr.addstr(21, 35, " QRZ   Cloudlog   ")
+    stdscr.addstr(21, 40, YorN(qrzsession), highlightBonus(qrzsession))
+    stdscr.addstr(21, 51, YorN(cloudLogOn), highlightBonus(cloudLogOn))
     stdscr.addstr(23, 0, "Band       Freq             Mode   ")
     stdscr.addstr(23, 5, strband.rjust(5), curses.A_REVERSE)
     stdscr.addstr(23, 16, strfreq, curses.A_REVERSE)
@@ -1316,10 +1351,13 @@ def setclass(c):
 
 
 def setsection(s):
-    global mysection
-    mysection = str(s)
-    writepreferences()
-    statusline()
+    global mysection, sections
+    if s and str(s) in validSections:
+        mysection = str(s)
+        writepreferences()
+        statusline()
+    else:
+        setStatusMsg("Must be valid section")
 
 
 def setrigctrlhost(o):
@@ -1475,19 +1513,21 @@ def processcommand(cmd):
         sendRadio(cmd[:1], cmd[1:])
         return
     if cmd[:1] == "B":  # Change Band
-        if rigonline:
-            sendRadio(cmd[:1], cmd[1:])
-            return
+        if cmd[1:] and cmd[1:] in bands:
+            if rigonline:
+                sendRadio(cmd[:1], cmd[1:])
+                return
+            else:
+                setband(cmd[1:])
         else:
-            setband(cmd[1:])
+            setStatusMsg('Specify valid band')
         return
     if cmd[:1] == "M":  # Change Mode
         if rigonline == False:
             if cmd[1:] == "CW" or cmd[1:] == "PH" or cmd[1:] == "DI":
                 setmode(cmd[1:])
             else:
-                curses.flash()
-                curses.beep()
+                setStatusMsg("Must be CW, DI, PH")
         else:
             if (
                 cmd[1:] == "USB"
@@ -1615,9 +1655,13 @@ def proc_key(key):
             return
         if hiscall == "" or hisclass == "" or hissection == "":
             return
-        contact = (hiscall, hisclass, hissection, band, mode, int(power))
-        log_contact(contact)
-        clearentry()
+        isCall = re.compile("[A-Z]{1,3}[0-9]{1,3}[A-z]{1,3}(/[A-Z0-9]{1,3})?")
+        if re.match(isCall, hiscall):
+            contact = (hiscall, hisclass, hissection, band, mode, int(power))
+            log_contact(contact)
+            clearentry()
+        else:
+            setStatusMsg("Must be valid call sign")
         return
     elif key == Escape:
         clearentry()
@@ -1781,6 +1825,9 @@ def EditClickedQSO(line):
 
 
 def editQSO(q):
+    if q == False or q == "":
+        setStatusMsg("Must specify a contact number")
+        return
     global qsoew, qso, quit
     conn = sqlite3.connect(database)
     c = conn.cursor()
